+++ CHANGELOG.rst
@@ -15,6 +15,10 @@ file carefully before upgrading.
 * The command line option ``--no-cache`` has been renamed to
   ``--no-config-file``. The old name is retained as an alias for backwards
   compatibility
+* Added a --revision command line option. In the case of apply, this causes
+  the specified migration to be applied, plus any dependencies. In the case
+  of rollback, this removes the specified revision and any other migrations
+  that depend upon it.
 
 Version 4.2.5
 ~~~~~~~~~~~~~